Data Extraction Module/requirements.txt
Flask
pytesseract
Pillow
serverless-wsgi